fix: fix build with clang-18
Some uses of inline assembly for accessing floating point registers did
not enable the FP extension and were failing with clang-18.

The check in the linker script for the `hftest_enable` variable fails
under clang-18. However, hftest still compiles and passes without the
check, so it was removed.

Fix new warnings from `clang-tidy`.

Apply `make format`.

+++ src/dlog.c
@@ -393,7 +393,7 @@ uint64_t reinterpret_signed_int(enum format_length length, uint64_t value,
 		}
 		break;
 	case length64:
-		if ((int64_t)signed_value < 0) {
+		if (signed_value < 0) {
 			flags->neg = true;
 			signed_value = -signed_value;
 		}
